burner: move interface and mock into its own files
They do not really belong to the overall controller setup. From the
point of view of a user, it's more of a blackbox device to start and
stop the burning of gas and it's less interesting that it internally has
some form of control loop.
